ChatwithPDF_GenAI/requirements.txt
langchain
accelerate
tiktoken
pypdf
python-dotenv
chromadb
openai